+++ src/libraries/vup/Shader/Compute_shader.h
@@ -18,10 +18,10 @@ namespace vup
                                 const std::vector<Shader_define>& defines = {},
                                 gl::introspection introspection_flag = gl::introspection::basic);
         void reload() override;
-        void run(float x = 1, float y = 1, float z = 1) const;
-        void run_with_barrier(float x = 1, float y = 1, float z = 1,
+        void run(unsigned int x = 1, unsigned int y = 1, unsigned int z = 1) const;
+        void run_with_barrier(unsigned int x = 1, unsigned int y = 1, unsigned int z = 1,
                               GLbitfield barriers = GL_SHADER_STORAGE_BARRIER_BIT) const;
-        void run_workgroups(GLuint x = 1, GLuint y = 1, GLuint z = 1) const;
+        void run_workgroups(unsigned int x = 1, unsigned int y = 1, unsigned int z = 1) const;
         std::array<GLint, 3> get_workgroup_size() const;
         GLint get_workgroup_size_x() const;
         GLint get_workgroup_size_y() const;
